+++ c#/imagePipeline/src/Db/ImageQrCode.cs
@@ -4,13 +4,13 @@ public class ImageQrCode
 {
     public uint ImageId { get; set; }
     public uint FrameIndex { get; set; }
-    public float Point1X { get; set; }
-    public float Point1Y { get; set; }
-    public float Point2X { get; set; }
-    public float Point2Y { get; set; }
-    public float Point3X { get; set; }
-    public float Point3Y { get; set; }
-    public float Point4X { get; set; }
-    public float Point4Y { get; set; }
+    public short Point1X { get; set; }
+    public short Point1Y { get; set; }
+    public short Point2X { get; set; }
+    public short Point2Y { get; set; }
+    public short Point3X { get; set; }
+    public short Point3Y { get; set; }
+    public short Point4X { get; set; }
+    public short Point4Y { get; set; }
     public required string Text { get; set; }
 }
